+++ world/src/stringf.h
@@ -2,14 +2,15 @@
 #define __stringf_h__
 
 #include <string>
+#include <ctype.h>
 
-std::string toHex(unsigned int a);
-std::string toHex(unsigned long long a);
-std::string toString(unsigned long long a);
-std::string toString(unsigned int a);
-std::string toString(long long a);
+std::string toHex(uint32_t a);
+std::string toHex(uint64_t a);
+std::string toString(uint32_t a);
+std::string toString(uint64_t a);
+std::string toString(int64_t a);
 std::string operator+(const char *s, const std::string& s2);
-std::string toString(int a);
+std::string toString(int32_t a);
 std::string toString(float a);
 int toInt(const std::string &str);
 float toFloat(const std::string &str);
